# Add benchmarks to monitor lookup hit/miss performances depending on load factor