Current implementation incurs a performance penalty as table gets full, in particular for lookup miss.
This is a good way to quantify and monitor this behavior so we can try improving it in the future.

Typical results look like this:
```
// FoldHashMap, RandomKeys
test loadfactor_lookup_14500             ... bench:       3,505.05 ns/iter (+/- 37.71)
test loadfactor_lookup_16500             ... bench:       3,474.36 ns/iter (+/- 11.88)
test loadfactor_lookup_18500             ... bench:       3,482.29 ns/iter (+/- 14.58)
test loadfactor_lookup_20500             ... bench:       3,478.73 ns/iter (+/- 20.59)
test loadfactor_lookup_22500             ... bench:       3,416.84 ns/iter (+/- 94.83)
test loadfactor_lookup_24500             ... bench:       3,450.97 ns/iter (+/- 82.86)
test loadfactor_lookup_26500             ... bench:       3,493.92 ns/iter (+/- 37.22)
test loadfactor_lookup_28500             ... bench:       3,484.17 ns/iter (+/- 24.00)

test loadfactor_lookup_fail_14500        ... bench:       2,900.72 ns/iter (+/- 28.27)
test loadfactor_lookup_fail_16500        ... bench:       3,009.21 ns/iter (+/- 23.97)
test loadfactor_lookup_fail_18500        ... bench:       3,157.28 ns/iter (+/- 24.10)
test loadfactor_lookup_fail_20500        ... bench:       3,480.48 ns/iter (+/- 27.06)
test loadfactor_lookup_fail_22500        ... bench:       4,048.19 ns/iter (+/- 47.99)
test loadfactor_lookup_fail_24500        ... bench:       5,265.12 ns/iter (+/- 65.23)
test loadfactor_lookup_fail_26500        ... bench:       7,494.26 ns/iter (+/- 90.56)
test loadfactor_lookup_fail_28500        ... bench:      11,593.97 ns/iter (+/- 108.62)
```