@gaujay 2025-08-21
Removed clippy fix. Ready to merge.